+++ src/GafferSceneUIModule/SceneInspectorBinding.cpp
@@ -710,21 +710,27 @@ vector<InternedString> alphabeticallySortedKeys( const T &container )
 	return result;
 }
 
-const boost::container::flat_map<string, InternedString> g_attributeCategories = {
-	{ "ai:*", "Arnold" },
-	{ "dl:*", "3Delight" },
-	{ "cycles:*", "Cycles" },
-	{ "ri:*", "RenderMan" },
-	{ "gl:*", "OpenGL" },
-	{ "usd:*", "USD" },
-	{ "user:*", "User" },
+boost::container::flat_map<string, InternedString> g_attributeCategories;
+
+void registerAttributeCategory( InternedString category, const IECore::StringAlgo::MatchPattern &pattern )
+{
+	g_attributeCategories.insert( { pattern, category } );
+}
+
+void deregisterAttributeCategory( InternedString category )
+{
+	for( auto it = g_attributeCategories.begin(); it != g_attributeCategories.end(); )
 	{
-		"scene:visible doubleSided render:* gaffer:* "
-		"linkedLights shadowedLights filteredLights "
-		"surface displacement volume light",
-		"Standard"
+		if( it->second == category )
+		{
+			it = g_attributeCategories.erase( it );
+		}
+		else
+		{
+			it++;
+		}
 	}
-};
+}
 
 const InternedString g_other( "Other" );
 
@@ -1348,16 +1354,27 @@ const InspectorTree::Registration g_subdivisionInspectionRegistration( { "Locati
 // Option Inspectors
 // =================
 
-const boost::container::flat_map<string, InternedString> g_optionCategories = {
-	{ "ai:*", "Arnold" },
-	{ "dl:*", "3Delight" },
-	{ "cycles:*", "Cycles" },
-	{ "ri:*", "RenderMan" },
-	{ "gl:*", "OpenGL" },
-	{ "usd:*", "USD" },
-	{ "user:*", "User" },
-	{ "render:* sampleMotion", "Standard" },
-};
+boost::container::flat_map<string, InternedString> g_optionCategories;
+
+void registerOptionCategory( InternedString category, const IECore::StringAlgo::MatchPattern &pattern )
+{
+	g_optionCategories.insert( { pattern, category } );
+}
+
+void deregisterOptionCategory( InternedString category )
+{
+	for( auto it = g_optionCategories.begin(); it != g_optionCategories.end(); )
+	{
+		if( it->second == category )
+		{
+			it = g_optionCategories.erase( it );
+		}
+		else
+		{
+			it++;
+		}
+	}
+}
 
 const std::string g_optionPrefix( "option:" );
 const std::string g_attributePrefix( "attribute:" );
@@ -1700,5 +1717,11 @@ void GafferSceneUIModule::bindSceneInspector()
 			.value( "B", InspectorDiffColumn::DiffContext::B )
 		;
 	}
+
+	def( "registerAttributeCategory", &registerAttributeCategory, ( boost::python::arg( "category" ), boost::python::arg( "pattern" ) ) );
+	def( "deregisterAttributeCategory", &deregisterAttributeCategory, ( boost::python::arg( "category" ) ) );
+
+	def( "registerOptionCategory", &registerOptionCategory, ( boost::python::arg( "category" ), boost::python::arg( "category" ) ) );
+	def( "deregisterOptionCategory", &deregisterOptionCategory, ( boost::python::arg( "category" ) ) );
 
 }