+++ startup/gui/sceneInspector.py
@@ -1,0 +1,67 @@
+##########################################################################
+#
+#  Copyright (c) 2025, Cinesite VFX Ltd. All rights reserved.
+#
+#  Redistribution and use in source and binary forms, with or without
+#  modification, are permitted provided that the following conditions are
+#  met:
+#
+#      * Redistributions of source code must retain the above
+#        copyright notice, this list of conditions and the following
+#        disclaimer.
+#
+#      * Redistributions in binary form must reproduce the above
+#        copyright notice, this list of conditions and the following
+#        disclaimer in the documentation and/or other materials provided with
+#        the distribution.
+#
+#      * Neither the name of John Haddon nor the names of
+#        any other contributors to this software may be used to endorse or
+#        promote products derived from this software without specific prior
+#        written permission.
+#
+#  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+#  IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+#  THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+#  PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+#  CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+#  EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+#  PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+#  PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+#  LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+#  NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+#  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#
+##########################################################################
+
+import GafferSceneUI
+
+## \todo Investigate alternatives to these manual registrations. Perhaps we could register
+# category metadata for each attribute/option and use it directly in the SceneInspector instead?
+
+for category, pattern in {
+	"Arnold" : "ai:*",
+	"3Delight" : "dl:*",
+	"Cycles" : "cycles:*",
+	"RenderMan" : "ri:*",
+	"OpenGL" : "gl:*",
+	"USD" : "usd:*",
+	"User" : "user:*",
+	"Standard" :
+		"scene:visible doubleSided render:* gaffer:* "
+		"linkedLights shadowedLights filteredLights "
+		"surface displacement volume light",
+}.items() :
+	GafferSceneUI.SceneInspector.registerAttributeCategory( category, pattern )
+
+for category, pattern in {
+	"Arnold" : "ai:*",
+	"3Delight" : "dl:*",
+	"Cycles" : "cycles:*",
+	"RenderMan" : "ri:*",
+	"OpenGL" : "gl:*",
+	"USD" : "usd:*",
+	"User" : "user:*",
+	"Standard" : "render:* sampleMotion",
+}.items() :
+	GafferSceneUI.SceneInspector.registerOptionCategory( category, pattern )